submit abc169_c

+++ abc/169/c.cpp
@@ -18,12 +18,12 @@ int main() {
   // ans = (int64_t)b;
 
   int64_t a, ans;
-  double b;
+  long double b;
   cin >> a >> b;
   // cerr << (int)(b * 100) << '\n';
   // cerr << floor(b * 100) << '\n';
   // cerr << round(b * 100) << '\n';
-  ans = (a * round(b * 100)) / 100;
+  ans = (a * (int64_t)round(b * 100)) / 100;
 
   cout << ans << '\n';
